Tweaks to make it work with latest Pandoc
Somehow, it inserts a bunch of {.title-ref} things into the Markdown
after conversion from .rst, so strip those out.  Also turn off
rewrapping of the changelog for more consistent results.  This is all a
bit of a workaround right now; there might be a more official way to do
this.